[dhcpv4] move default to main directory
